requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
confluent-kafka==2.3.0
